requirements.txt
Flask==3.0.3
pandas==1.3.3
scikit-learn==1.5.0
python-dotenv==1.0.1
pymongo==4.8.0
pytest==8.2.2
Flask-Testing==0.8.1
Flask-PyMongo==2.3.0
imbalanced-learn==0.12.3
numpy==1.21.2

